apps/backend/requirements.txt
uvicorn[standard]
requests

python-dotenv

fastapi
fastapi-azure-auth

google-cloud-documentai
google-generativeai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

langchain-text-splitters
langchain-google-genai
langchain-postgres
langchain-community
langchain-openai
langchain-core
langchain
langgraph

protobuf

pydantic
pydantic-settings
pydantic[email]

llama-cloud
llama-cloud-services
llama-index-embeddings-openai

PyJWT[crypto]

sqlalchemy
asyncpg
psycopg
psycopg2-binary
sqlacodegen[pgvector]
pgvector

python-multipart

pandas

google-genai
openai

sqlacodegen

pytesseract
pillow

strands-agents 
strands-agents-tools

pypdf

chromadb